Speedup & completion of download on errors. (#30)
* Replaces errors by warnings during image download.

If an error occurs during downloading an image from a set, a warning instead of
an error is emitted. This allows the rest of the set being downloaded.

* Avoids unnecessary Flickr accesses.

When image overrides are turned off, the check of an existing file should be
done before getting the available image sizes from Flickr. This avoids
unnecessary Flickr accesses and speeds up the process.